backend/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
apscheduler>=3.10.0
playwright>=1.44.0
beautifulsoup4>=4.12.0
openai>=1.30.0
python-docx>=1.1.0
reportlab>=4.0.0
python-multipart>=0.0.9
